// MangoTech/Hospital/src/app/treatment/customers/model/customer.model.ts
export class CustomerModel {
  id: string;
  userId: string;
  username: string;
  job: string;
  address: string;
  avatar: string;
  servicesRequest: string;
  bookingDate: number;
  incomingDate: number;
  modified: number;
  modifier: string;
  customerType: number;
  services: string;
  customerName: string;
  phone: string;
  email: string;
  facebook: string;
  approach: number;
  gender: number;
  birthDate: number;
  created: number;
  creator: string;
  status: number;
  removed: boolean;
}